Eliminate warnings so I can see actual build output

+++ main.cpp
@@ -19,7 +19,7 @@ std::vector<void*> Text;
 void* Enemy;
 void* U;
 void* bull;
-int UX = 400, UY = 550;
+float UX = 400, UY = 550;
 
 void Initialize()
 {
@@ -70,13 +70,13 @@ void Game()
 			int xo = 0, yo = 0;
 			int n1 = time + n * n + n * n * n;
 			int n2 = time + n + n * n + n * n * n * 3;
-			if (((n1 >> 6) & 0x7) == 0x7)xo += (1 - cos((n1 & 0x7f) / 64.0f * 2.f * 3.141592)) * (20 + ((n * n) % 9));
-			if (((n1 >> 6) & 0x7) == 0x7)yo += (sin((n1 & 0x7f) / 64.0f * 2.f * 3.141592)) * (20 + ((n * n) % 9));
-			if (((n2 >> 8) & 0xf) == 0xf)yo += (1 - cos((n2 & 0xff) / 256.0f * 2.f * 3.141592)) * (150 + ((n * n) % 9));
-			DrawSprite(Enemy, x[n] + xo, y[n] + yo, (10 + ((n) % 17)), (10 + ((n) % 17)), 0, 0xffffffff);
+			if (((n1 >> 6) & 0x7) == 0x7)xo += (int)((1 - cos((n1 & 0x7f) / 64.0f * 2.f * 3.141592)) * (20 + ((n * n) % 9)));
+			if (((n1 >> 6) & 0x7) == 0x7)yo += (int)((sin((n1 & 0x7f) / 64.0f * 2.f * 3.141592)) * (20 + ((n * n) % 9)));
+			if (((n2 >> 8) & 0xf) == 0xf)yo += (int)((1 - cos((n2 & 0xff) / 256.0f * 2.f * 3.141592)) * (150 + ((n * n) % 9)));
+			DrawSprite(Enemy, (float)x[n] + xo, (float)y[n] + yo, (float)(10 + ((n) % 17)), (float)(10 + ((n) % 17)), 0, 0xffffffff);
 		}
 
-		DrawSprite(U, UX += IsKeyDown(VK_LEFT) ? -7 : IsKeyDown(VK_RIGHT) ? 7 : 0, UY, 50, 50, 3.141592 + sin(time * 0.1) * 0.1, 0xffffffff);
+		DrawSprite(U, UX += IsKeyDown(VK_LEFT) ? -7 : IsKeyDown(VK_RIGHT) ? 7 : 0, UY, 50, 50, (float)(3.141592 + sin(time * 0.1) * 0.1), 0xffffffff);
 
 		// FIRE
 		static int b = 0;
@@ -85,8 +85,8 @@ void Game()
 		if (!IsKeyDown(VK_SPACE)) count = 0;
 		if (IsKeyDown(VK_SPACE) && count == 0)
 		{ 
-			bullets[b].BX = UX; 
-			bullets[b].BY = UY; 
+			bullets[b].BX = (float)UX;
+			bullets[b].BY = (float)UY;
 			b = (b + 1) % 10; 
 			count = 15;
 		}
@@ -97,10 +97,10 @@ void Game()
 			DrawSprite(bull, bullets[n].BX, bullets[n].BY -= 4, 10, 10, bullets[n].BA += 0.1f, 0xffffffff);
 		}
 
-		for (int n = 0; n < strlen("space invaders"); ++n)
+		for (unsigned int n = 0; n < strlen("space invaders"); ++n)
 		{
 			if (n != 5)
-				DrawSprite(Text[n], n * 40 + 150, 30, 20, 20, sin(time * 0.1) * n * 0.01);
+				DrawSprite(Text[n], (float)(n * 40 + 150), 30, 20, 20, (float)(sin(time * 0.1) * n * 0.01));
 		}
 
 		Flip();
